v0.0.6 2022-05-07
utils调整
1. jwt.create_token支持yaml格式调整

+++ src/lesoon_common/wrappers/plugins.py
@@ -234,7 +234,8 @@ class Bootstrap:
                 app.logger.info('文件锁已被占领, 等待其他进程重载配置文件...')
         while not os.path.lexists(success_filename):
             continue
-        os.remove(success_filename)
+        if os.path.exists(success_filename):
+            os.remove(success_filename)
         app.__class__.config_path = config_filename
 
     def reload_config_from_configmap(self, parser: configparser.ConfigParser,
